requirements.txt
langchain==0.1.12
langchain-groq==0.0.1
chromadb==0.4.24
pandas==2.2.1
python-dotenv==1.0.1
streamlit==1.32.2 
sentence-transformers